Exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -50,9 +50,24 @@ blogsRouter.get('/', async (request, response) => {
   })
 
   blogsRouter.delete('/:id', async (request, response) => {
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if(!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid '})
+    }
+    const user = await User.findById(decodedToken.id)
+    const userId = user.id
+
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog.user.toString() === userId.toString() ) {
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
-  })
+ 
+    }
+
+     })
 
   blogsRouter.put('/:id', async (request, response) => {
     const body = request.body
